move ke.h
