+++ ydb/services/persqueue_v1/actors/events.h
@@ -5,7 +5,7 @@
 #include <ydb/core/base/events.h>
 #include <ydb/core/grpc_services/rpc_calls_topic.h>
 #include <ydb/core/protos/pqconfig.pb.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/public/key.h>
 #include <ydb/core/persqueue/percentile_counter.h>
 #include <ydb/core/tx/scheme_board/events.h>
